Add coi::base class
The coi::base class has been created to
replace logic that had previously been
stored in both the openstack parent node
in grizzly-manifests site.pp as well as
the coi::profiles::openstack::base class.

It is intended to serve as a class that hold
logic that should be applied to all opensack
nodes. The majority of its logic is currently
for configuring repos.