Adds Support for Network Interface Options
Previously, the class did not support managing network interface
options.  This is required for nodes that have network interfaces
configured for bridged mode (i.e. control/compute nodes).

The patch adds the $options parameter which defaults to undef for
backwards compatibility and for network interface options not to
be configured.