Make coe::nexus work with Havana
The coe::nexus class was still referring to OpenStack
Networking as Quantum.  Since the name was changed to
Neutron in Havana, we need to follow suit.  This patch
changes all references from Quantum to Neutron.

Additionally, the configuration files for the Cisco plugins
have changed in Havana.  This patch provides a new template
and appropriate logic to populate it.

Partial-Bug: #1276168